Create /etc/containers and install iptables
This is primarily for Ubuntu. For whatever reason the Ubuntu packages
for podman do a poor job of pulling in dependencies and ensuring their
expected state. Containers shake this out because the Ubuntu
container does not include many things that the podman team is taking
for granted. We add these things so that our containers and any other
similiarly minimal Linux distros will work with podman after
installed.